requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-socketio>=5.11.0
httpx>=0.28.0
Pillow>=12.0.0
qrcode[pil]>=7.4
aiofiles>=24.0.0
pyyaml>=6.0
